refactor(core): core 모듈에서 infra 모듈로 job과 prompt를 분리한다.
