+++ resumarble-infrastructure/src/main/kotlin/resumarble/infrastructure/job/JobPersistenceAdapter.kt
@@ -1,0 +1,13 @@
+package resumarble.infrastructure.job
+
+import resumarble.core.domain.job.domain.Job
+import resumarble.core.domain.job.infrastructure.JobRepository
+import resumarble.infrastructure.annotation.Adapter
+import resumarble.infrastructure.job.entity.JobJpaRepository
+
+@Adapter
+class JobPersistenceAdapter(
+    private val jobJpaRepository: JobJpaRepository
+) : JobRepository {
+    override fun findAll(): List<Job> = jobJpaRepository.findAll()
+}
