+++ resumarble-infrastructure/src/main/kotlin/resumarble/infrastructure/job/entity/JobJpaRepository.kt
@@ -1,4 +1,4 @@
-package resumarble.core.domain.job.infrastructure
+package resumarble.infrastructure.job.entity
 
 import org.springframework.data.jpa.repository.JpaRepository
 import resumarble.core.domain.job.domain.Job
