+++ resumarble-infrastructure/src/main/kotlin/resumarble/infrastructure/prompt/entity/PromptJpaRepository.kt
@@ -1,4 +1,4 @@
-package resumarble.core.domain.prompt.infrastructure
+package resumarble.infrastructure.prompt.entity
 
 import org.springframework.data.jpa.repository.JpaRepository
 import resumarble.core.domain.prompt.domain.Prompt
